Update README.md
This list will be like the sequence class that we created for Project Two, but it will be implemented differently, and the iterator access will be done as an external iterator. This list will also be implemented as a template, allowing it to hold any type of data. This will be a universal list.

Our list will be a doubly linked list. In a doubly linked list, each node has a pointer pointing to the next node and another one pointing to the previous node. (The “node” after the last item is NULL and the “node” previous to the first one is NULL as well.) This will require the definition of a different node class from what you used in project 3. So, you should start your project by making a Dnode class for nodes appropriate for use in a doubly linked list. Make this class template so it can hold any kind of data. You will need functions to access next, previous, and data as well as functions to set_next, set_previous and set_data. You should provide a constructor that uses default arguments that will set the pointers to NULL.